+++ stac/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.8.0] - 2024-08-12
+
 ### Added
 
 - `Version` enum ([#293](https://github.com/stac-utils/stac-rs/pull/293))
@@ -320,7 +322,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 Initial release.
 
-[Unreleased]: https://github.com/stac-utils/stac-rs/compare/stac-v0.7.2...main
+[Unreleased]: https://github.com/stac-utils/stac-rs/compare/stac-v0.8.0...main
+[0.8.0]: https://github.com/stac-utils/stac-rs/compare/stac-v0.7.2...stac-v0.8.0
 [0.7.2]: https://github.com/stac-utils/stac-rs/compare/stac-v0.7.1...stac-v0.7.2
 [0.7.1]: https://github.com/stac-utils/stac-rs/compare/stac-v0.7.0...stac-v0.7.1
 [0.7.0]: https://github.com/stac-utils/stac-rs/compare/stac-v0.6.0...stac-v0.7.0